# Use KernelAbstractions.jl for upsample kernels