--- a/ext/NNlibCUDA/src/NNlibCUDA.jl
+++ b/ext/NNlibCUDA/src/NNlibCUDA.jl
@@ -6,7 +6,6 @@ using Random, Statistics
 
 const IntOrIntTuple = Union{Integer, NTuple{N,<:Integer} where N}
 
-include("upsample.jl")
 include("sampling.jl")
 include("activations.jl")
 include("batchedadjtrans.jl")