--- a/src/NNlib.jl
+++ b/src/NNlib.jl
@@ -1,19 +1,20 @@
 module NNlib
 
-using Pkg
-using Requires
-using ChainRulesCore
 import ChainRulesCore: rrule
+
 using Base.Broadcast: broadcasted
 using Base.Threads
-using Random
-using Statistics
-using Statistics: mean
-using LinearAlgebra
-using LinearAlgebra: BlasFloat, Transpose, Adjoint, AdjOrTransAbsMat
-using LinearAlgebra.BLAS: BlasInt, @blasfunc
+using ChainRulesCore
 using KernelAbstractions
 using KernelAbstractions: @atomic
+using LinearAlgebra
+using LinearAlgebra.BLAS: @blasfunc, BlasInt
+using LinearAlgebra: AdjOrTransAbsMat, Adjoint, BlasFloat, Transpose
+using Pkg
+using Random
+using Requires
+using Statistics
+using Statistics: mean
 
 const libblas = Base.libblas_name