… Project.toml
… ext/NNlibCUDA/src/upsample.jl
… ext/NNlibCUDA/test/runtests.jl
… ext/NNlibCUDA/test/upsample.jl
… test/Project.toml
… test/test_utils.jl
… test/upsample.jl