Ref #479.

- Migrate upsample kernels to KernelAbstractions.jl (implementation migrated from NNlibCUDA).
  Successfully tested on `CPU`, `ROCBackend`, `CUDABackend`.
  From the user perspective, there are no API changes (except `align_corners` option).
  Backend is selected automatically using `KernelAbstractions.get_backend(x)`.

- Introduce NNlib test suite, which accepts backend as an argument and runs tests for respective device.
  For now, it contains only upsample tests, but as we move other kernels to KernelAbstraction, it makes sense to move the rest of the tests to it as well.
  This way, we don't need to rewrite tests for each backend.

  When backend is
  - `CPU`, gradtest checks Zygote gradient against FiniteDifferences (as was before).
  - **not** `CPU`, gradtest checks agains `CPU`.
  This is because AMDGPU.jl fails with FiniteDifferences.jl.
  Still, fow now this is fine, because test suite for `CPU` runs first.
```julia
AssertionError: ROCArray only supports bits types
  Stacktrace:
    [1] ROCVector{Tuple{Vector{Float32}, FiniteDifferences.var"#Real_from_vec#20"}}(buf::AMDGPU.Runtime.Mem.Buffer, dims::Tuple{Int64}; offset::Int64)
    ...
```
- Add `align_corners` option (copied from NNlibCUDA.jl, but now supports CPU as well).
- Remove upsample from `ext/NNlibCUDA`.

**Side note:** this PR is based on #483 since it introduces separate `Project.toml` for tests. 

- [x] Add benchmarks comparing performance before & after.

### PR Checklist

- [x] Tests are added
- [ ] Documentation, if applicable