@pxl-th 2023-04-06
Benchmarks

**CPU:**
`TODO`

**CUDA:**

- Before:
```julia
julia> using CUDA, Zygote, BenchmarkTools, NNlib, NNlibCUDA

julia> x = CUDA.rand(Float32, 128, 128, 128, 16);

julia> @benchmark CUDA.@sync NNlib.upsample_bilinear(x, (2, 2))
BenchmarkTools.Trial: 951 samples with 1 evaluation.
 Range (min … max):  4.072 ms … 17.352 ms  ┊ GC (min … max): 0.00% … 9.22%
 Time  (median):     4.924 ms              ┊ GC (median):    0.00%
 Time  (mean ± σ):   5.193 ms ±  1.168 ms  ┊ GC (mean ± σ):  4.06% ± 7.58%

    ▁▂▃▆▆▅▆█▇▃                                                
  ▆▄██████████▆▇▁▁▅▄▁▄▄▁▁▁▁▁▁▁▄▁▁▁▄▅▁▄▅▅▆▅▇▄▆█▇█▅▄▆▅▆▇▆▆▅▅▆▆ █
  4.07 ms      Histogram: log(frequency) by time     9.38 ms <

 Memory estimate: 3.67 KiB, allocs estimate: 63.

julia> @benchmark CUDA.@sync Zygote.gradient(x -> sum(NNlib.upsample_bilinear(x, (2, 2))), x)
BenchmarkTools.Trial: 189 samples with 1 evaluation.
 Range (min … max):  23.853 ms … 31.900 ms  ┊ GC (min … max): 0.00% … 11.43%
 Time  (median):     25.381 ms              ┊ GC (median):    0.00%
 Time  (mean ± σ):   26.413 ms ±  2.252 ms  ┊ GC (mean ± σ):  2.73% ±  4.30%

          ▃█▇▅▃▁                                               
  ▃▃▃▅▃▇█▇██████▅▄▄▃▃▁▁▃▃▁▁▁▁▃▄▃▃▁▁▃▃▁▁▃▁▁▃▁▃▃▃▄▃▃▃▃▃▃▄▃▃▅▃▄▃ ▃
  23.9 ms         Histogram: frequency by time        31.7 ms <

 Memory estimate: 18.97 KiB, allocs estimate: 405.
```
- This PR:
```julia
julia> using CUDA, Zygote, BenchmarkTools, NNlib

julia> x = CUDA.rand(Float32, 128, 128, 128, 16);

julia> @benchmark CUDA.@sync NNlib.upsample_bilinear(x, (2, 2))
BenchmarkTools.Trial: 1080 samples with 1 evaluation.
 Range (min … max):  4.039 ms …  17.729 ms  ┊ GC (min … max): 0.00% … 0.00%
 Time  (median):     4.426 ms               ┊ GC (median):    0.00%
 Time  (mean ± σ):   4.611 ms ± 728.586 μs  ┊ GC (mean ± σ):  1.39% ± 3.29%

  ▂▄▆█▆▆▅▆▅▅▄▅▄▅▂▃▂▂▁                                         ▁
  ███████████████████▇▅▄▄▄▆▄▆▄▆▁▆▆▇▇▆▆▆▅▆▄▄▆▆▄▆▄▄▆▆▅▁▆▆▁▅▁▆▆▇ █
  4.04 ms      Histogram: log(frequency) by time      7.05 ms <

 Memory estimate: 6.52 KiB, allocs estimate: 128.

julia> @benchmark CUDA.@sync Zygote.gradient(x -> sum(NNlib.upsample_bilinear(x, (2, 2))), x)
BenchmarkTools.Trial: 199 samples with 1 evaluation.
 Range (min … max):  22.150 ms … 30.296 ms  ┊ GC (min … max): 0.00% … 5.63%
 Time  (median):     24.243 ms              ┊ GC (median):    0.00%
 Time  (mean ± σ):   25.103 ms ±  1.843 ms  ┊ GC (mean ± σ):  2.07% ± 3.29%

              ▇▄▄█▂█                                           
  ▃▁▁▁▁▁▁▁▁▁▃████████▆▅▅▃▃▃▃▁▁▁▁▁▁▁▁▃▁▁▁▁▃▁▄▃▅▃▅▅▄▃▃▄▅▄▅▃▃▃▃▄ ▃
  22.2 ms         Histogram: frequency by time        29.3 ms <

 Memory estimate: 25.81 KiB, allocs estimate: 560.
```