@CarloLucibello 2023-04-10
Instead of wrapping tests within functions, can we define at the global level a `device(x)` (based on the available backend) and `cpu(x)` function and maintain the current style?