@pxl-th 2023-04-10
> Instead of wrapping tests within functions, can we define at the global level a `device(x)` (based on the available backend) and `cpu(x)` function and maintain the current style?

For `device(x)` to make it global we would also have to switch some global `DEVICE` variable, to be able to call `nnlib_testsuite(backend)` multiple times during the tests.

```julia
DEVICE = CPU
nnlib_testsuite(CPU) # device(x) = adapt(CPU, x)
...
DEVICE = ROCBackend
nnlib_testsuite(ROCBackend) # device(x) = adapt(ROCBackend, x)
```

If this is preferred, I can make it this way.
But that would make `device(x)` type unstable, no?

Alternatively, we can leave tests as functions and define `device(x)` there.

For I've added `cpu(x)` as global function and `device(x)` is defined in `upsample_testsuite(x)` function.