@ToucheSir 2023-04-10
Sorry, I forgot to ask how perf compares on CPU with varying numbers of threads. We should probably have a separate PR adding upsampling to the benchmarks suite.