@pxl-th 2023-04-10
I've optimized CPU kernels a bit more, they are now even faster.

**`--threads=1`:**

- Before:
```julia
julia> @benchmark NNlib.upsample_bilinear(x, (2, 2))
BenchmarkTools.Trial: 14 samples with 1 evaluation.
 Range (min … max):  364.796 ms … 380.752 ms  ┊ GC (min … max): 0.06% … 2.46%
 Time  (median):     374.844 ms               ┊ GC (median):    2.49%
 Time  (mean ± σ):   374.699 ms ±   3.776 ms  ┊ GC (mean ± σ):  2.14% ± 0.89%

                                      █ ▄                        
  ▆▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▆▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁█▁█▁▁▁▁▁▆▆▆▁▁▁▁▆▁▁▁▁▁▁▁▁▆ ▁
  365 ms           Histogram: frequency by time          381 ms <

 Memory estimate: 512.00 MiB, allocs estimate: 10.

julia> @benchmark Zygote.gradient(x -> sum(NNlib.upsample_bilinear(x, (2, 2))), x)
BenchmarkTools.Trial: 7 samples with 1 evaluation.
 Range (min … max):  780.481 ms … 801.564 ms  ┊ GC (min … max): 0.13% … 2.20%
 Time  (median):     799.364 ms               ┊ GC (median):    2.09%
 Time  (mean ± σ):   796.938 ms ±   7.326 ms  ┊ GC (mean ± σ):  1.81% ± 0.75%

  ▁                                                   ▁▁█  ▁  ▁  
  █▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁███▁▁█▁▁█ ▁
  780 ms           Histogram: frequency by time          802 ms <

 Memory estimate: 768.01 MiB, allocs estimate: 238.
```
- This PR:

```julia
julia> @benchmark NNlib.upsample_bilinear(x, (2, 2))
BenchmarkTools.Trial: 15 samples with 1 evaluation.
 Range (min … max):  342.028 ms … 352.197 ms  ┊ GC (min … max): 0.07% … 2.85%
 Time  (median):     351.279 ms               ┊ GC (median):    2.74%
 Time  (mean ± σ):   350.428 ms ±   2.765 ms  ┊ GC (mean ± σ):  2.40% ± 0.95%

                                                        █▄▁ ▄    
  ▆▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▆▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁███▁█▁▆ ▁
  342 ms           Histogram: frequency by time          352 ms <

 Memory estimate: 512.00 MiB, allocs estimate: 13.

julia> @benchmark Zygote.gradient(x -> sum(NNlib.upsample_bilinear(x, (2, 2))), x)
BenchmarkTools.Trial: 7 samples with 1 evaluation.
 Range (min … max):  752.479 ms … 768.879 ms  ┊ GC (min … max): 0.15% … 2.22%
 Time  (median):     766.059 ms               ┊ GC (median):    2.17%
 Time  (mean ± σ):   764.923 ms ±   5.610 ms  ┊ GC (mean ± σ):  1.86% ± 0.76%

  ▁                                                ▁█    ▁ ▁  ▁  
  █▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁██▁▁▁▁█▁█▁▁█ ▁
  752 ms           Histogram: frequency by time          769 ms <

 Memory estimate: 640.01 MiB, allocs estimate: 275.
```

**`--thread=16`:**

- Before:
```julia
julia> using BenchmarkTools, NNlib, Zygote

julia> x = rand(Float32, 128, 128, 128, 16);

julia> @benchmark NNlib.upsample_bilinear(x, (2, 2))
BenchmarkTools.Trial: 76 samples with 1 evaluation.
 Range (min … max):  47.665 ms … 74.961 ms  ┊ GC (min … max):  0.65% … 27.56%
 Time  (median):     67.239 ms              ┊ GC (median):    26.38%
 Time  (mean ± σ):   66.386 ms ±  4.412 ms  ┊ GC (mean ± σ):  27.19% ±  5.61%

                                      █▅        ▇▆             
  ▃▁▃▁▁▁▁▁▁▁▁▁▁▃▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▅██▆▄▁▅▁▁▁▄██▄▄▃▄▁▃▁▃▃▃▃ ▁
  47.7 ms         Histogram: frequency by time        73.9 ms <

 Memory estimate: 512.01 MiB, allocs estimate: 101.

julia> @benchmark Zygote.gradient(x -> sum(NNlib.upsample_bilinear(x, (2, 2))), x)
BenchmarkTools.Trial: 28 samples with 1 evaluation.
 Range (min … max):  154.072 ms … 191.380 ms  ┊ GC (min … max):  1.02% … 15.65%
 Time  (median):     180.883 ms               ┊ GC (median):    16.55%
 Time  (mean ± σ):   181.300 ms ±   6.365 ms  ┊ GC (mean ± σ):  15.91% ±  2.94%

                                            ██▁                  
  ▄▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▄▁▁▁▄███▇▇▁▄▁▁▁▁▄▇▁▁▁▁▁▇ ▁
  154 ms           Histogram: frequency by time          191 ms <

 Memory estimate: 768.03 MiB, allocs estimate: 422.
```

- This PR:
```julia
julia> @benchmark NNlib.upsample_bilinear(x, (2, 2))
BenchmarkTools.Trial: 87 samples with 1 evaluation.
 Range (min … max):  47.044 ms … 68.864 ms  ┊ GC (min … max):  0.80% … 16.62%
 Time  (median):     57.800 ms              ┊ GC (median):    19.52%
 Time  (mean ± σ):   57.725 ms ±  2.507 ms  ┊ GC (mean ± σ):  18.97% ±  3.56%

                                          ▁ ▇█▂▄▁▁ ▅           
  ▃▃▁▁▁▁▃▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▃▁▃▅▆▆█▅████████▆▆▃▁▁▅▅▁▃ ▁
  47 ms           Histogram: frequency by time        61.2 ms <

 Memory estimate: 512.01 MiB, allocs estimate: 167.

julia> @benchmark Zygote.gradient(x -> sum(NNlib.upsample_bilinear(x, (2, 2))), x)
BenchmarkTools.Trial: 30 samples with 1 evaluation.
 Range (min … max):  149.817 ms … 181.332 ms  ┊ GC (min … max):  0.93% … 10.90%
 Time  (median):     166.550 ms               ┊ GC (median):    11.34%
 Time  (mean ± σ):   166.666 ms ±   4.378 ms  ┊ GC (mean ± σ):  11.04% ±  1.96%

                               ▃  █▁                             
  ▄▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▄█▁▇██▇▄▄▁▄▁▄▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▄ ▁
  150 ms           Histogram: frequency by time          181 ms <

 Memory estimate: 640.04 MiB, allocs estimate: 584.
```